update now implementation to be browser friendly
